refactor(uv): reorganize the pep517_whl package by concern
Split rule.bzl following the one-rule-per-file convention py/private
already uses: pep517_whl.bzl (anyarch), pep517_native_whl.bzl (native +
cross), common.bzl (shared helpers and base attrs), with defs.bzl as the
package's single public surface. cc_layer.bzl and exec_transition.bzl
were already factored by concern and stay put.

The runtime tools move to tools/: build_helper.py (the main of every
generated build_tool py_binary) and memory_monitor.py, each still
following the package's constraints (single-file helper, conditional
py_library).

Mechanical fallout: the sdist_build BUILD template loads defs.bzl and
points at the tools/ labels (generated-repo snapshots regenerated in the
root and e2e/cases workspaces), test data labels and the
dirname/TEST_SRCDIR-walking path assumptions follow the move, and the CI
typecheck file lists pick up the new paths.
