+++ .github/workflows/ci-workflows.yaml
@@ -56,8 +56,8 @@ jobs:
                   py/tools/site_merge/site_merge.py
                   py/tools/unpack/unpack.py
                   uv/private/gazelle_manifest/generate.py
-                  uv/private/pep517_whl/build_helper.py
-                  uv/private/pep517_whl/memory_monitor.py
+                  uv/private/pep517_whl/tools/build_helper.py
+                  uv/private/pep517_whl/tools/memory_monitor.py
                   uv/private/py_entrypoint_binary/search.py
                 )
                 runtime311=(
