+++ e2e/cases/uv-sdist-mpicc/BUILD.bazel
@@ -9,7 +9,7 @@ load("@aspect_rules_py//py:defs.bzl", "py_test")
 py_test(
     name = "test",
     srcs = ["__test__.py"],
-    data = ["@aspect_rules_py//uv/private/pep517_whl:build_helper.py"],
+    data = ["@aspect_rules_py//uv/private/pep517_whl/tools:build_helper.py"],
     dep_group = "uv_sdist_mpicc",
     main = "__test__.py",
     python_version = "3.11",
