+++ uv/private/pep517_whl/common.bzl
@@ -1,0 +1,94 @@
+"""Shared helpers and attrs for the PEP 517 sdist-to-wheel rules."""
+
+load("@bazel_lib//lib:resource_sets.bzl", "resource_set_attr")
+load("//uv/private:source_built_wheel.bzl", "SourceBuiltWheelInfo")
+load(":exec_transition.bzl", "exec_transition")
+
+TARGET_EXEC_GROUP = "target"
+
+_INHERITED_PYTHON_ENV = (
+    "PYTHONHOME",
+    "PYTHONPATH",
+    "PYTHONPLATLIBDIR",
+)
+
+def wheel_providers(wheel_dir, console_scripts):
+    return [
+        DefaultInfo(files = depset([wheel_dir])),
+        SourceBuiltWheelInfo(console_scripts = tuple(console_scripts)),
+    ]
+
+def tool_files_to_run(ctx):
+    tool = ctx.attr.tool
+    if type(tool) == "list":
+        tool = tool[0]
+    return tool[DefaultInfo].files_to_run
+
+def common_env(ctx):
+    # pyproject_hooks copies the build process environment and launches its
+    # Python executable without -I:
+    # https://github.com/pypa/pyproject-hooks/blob/4b7c6d113fb89b755d762a88712c8a6873cddd47/src/pyproject_hooks/_impl.py#L70-L83
+    # https://github.com/pypa/pyproject-hooks/blob/4b7c6d113fb89b755d762a88712c8a6873cddd47/src/pyproject_hooks/_impl.py#L378-L396
+    # Host settings therefore must not replace that child's venv or stdlib.
+    # https://docs.python.org/3/using/cmdline.html#environment-variables
+    default_shell_env = {
+        key: value
+        for key, value in ctx.configuration.default_shell_env.items()
+        if key.upper() not in _INHERITED_PYTHON_ENV
+    }
+    return {
+        "SETUPTOOLS_SCM_PRETEND_VERSION": ctx.attr.version,
+        # Determinism: fix hash seed so dict/set iteration order is stable
+        "PYTHONHASHSEED": "0",
+        # Determinism: reproducible timestamps in archives
+        "SOURCE_DATE_EPOCH": "0",
+    } | default_shell_env
+
+def patch_args_and_inputs(ctx):
+    patch_args = []
+    patch_inputs = []
+    if ctx.attr.pre_build_patches:
+        patch_args.extend(["--patch-strip", str(ctx.attr.pre_build_patch_strip)])
+        for target in ctx.attr.pre_build_patches:
+            for f in target[DefaultInfo].files.to_list():
+                patch_args.extend(["--patch", f.path])
+                patch_inputs.append(f)
+    return patch_args, patch_inputs
+
+def memory_args(ctx):
+    return ["--monitor-memory"] if ctx.attr.monitor_memory else []
+
+_PATCH_ATTRS = {
+    "pre_build_patches": attr.label_list(
+        default = [],
+        allow_files = [".patch", ".diff"],
+        doc = "Patch files to apply to the extracted source before building.",
+    ),
+    "pre_build_patch_strip": attr.int(
+        default = 0,
+        doc = "Strip count for pre-build patches (-p flag to patch).",
+    ),
+}
+
+PEP517_WHL_ATTRS = {
+    "src": attr.label(allow_single_file = True),
+    # The wheel action runs in the "target" exec group; its frontend must run
+    # on the exec (host) platform but resolve its Python build dependencies
+    # for the *target* Python version. exec_transition pins --platforms to the
+    # host and resets the platform_libc/platform_version flags to host values
+    # (Starlark flags otherwise leak from the target configuration and break
+    # wheel selection for build deps). See exec_transition.bzl.
+    "tool": attr.label(executable = True, cfg = exec_transition),
+    "_allowlist_function_transition": attr.label(
+        default = "@bazel_tools//tools/allowlists/function_transition_allowlist",
+    ),
+    "version": attr.string(),
+    "console_scripts": attr.string_list(
+        doc = "Console scripts discovered from the source distribution's entry-point metadata.",
+    ),
+    "args": attr.string_list(default = ["--validate-anyarch"]),
+    "monitor_memory": attr.bool(
+        default = False,
+        doc = "Report approximate Linux process-tree RSS while building the wheel.",
+    ),
+} | _PATCH_ATTRS | resource_set_attr
