+++ uv/private/pep517_whl/tests/memory_monitor_test.py
@@ -11,7 +11,7 @@ import unittest
 from typing import Dict, List, Optional
 from unittest import mock
 
-from uv.private.pep517_whl import memory_monitor
+from uv.private.pep517_whl.tools import memory_monitor
 
 
 class MemoryMonitorTest(unittest.TestCase):
